requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
pyotp==2.9.0
python-dotenv==1.0.0
cryptography==41.0.4